add Makefile

+++ inc/Manager.hpp
@@ -6,7 +6,7 @@
 /*   By: aorji <aorji@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/04 18:43:46 by aorji             #+#    #+#             */
-/*   Updated: 2019/10/14 17:30:01 by aorji            ###   ########.fr       */
+/*   Updated: 2019/10/14 18:14:15 by aorji            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,14 +42,14 @@ private:
         std::vector<std::string> heuristics = {"-MisplacedTiles", "-ManhattanDistance", "-EuclideanDistance"};
         if (_ac == 2)
         {
-            for(int i = 0; i < heuristics.size(); ++i)
+            for(size_t i = 0; i < heuristics.size(); ++i)
                 if (heuristics[i] == _av[1])
                     _heuristic = static_cast<eHeuristic>(i + 1);
         }
         if (_ac != 2 || _heuristic == ERROR)
         {
             std::cout << "usage: " << heuristics.size() << " heuristics are available: " << std::endl;
-            for(int i = 0; i < heuristics.size(); ++i)
+            for(size_t i = 0; i < heuristics.size(); ++i)
                 std::cout << heuristics[i] << std::endl;
             throw InvalideHeuristic();
         }
